+++ BlueprintPurge/changelog.md
@@ -1,5 +1,9 @@
 ﻿# Changelog
 
+## [0.3.0]
+- fixed json syntax error
+- added basic syntax validation
+
 ## [0.2.0]
 - fixed rare case where a wrong id would be parsed, even tho it doesn't have any id
 - changed so affected list are cleared instead of setting them to null; renamed 'Null' field to 'IsList'